+++ spark-bigquery-dsv2/spark-3.1-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark31ReadFromQueryIntegrationTest.java
@@ -15,10 +15,6 @@
  */
 package com.google.cloud.spark.bigquery.integration;
 
-import org.junit.Ignore;
-
-// Not supported yet
-@Ignore
 public class Spark31ReadFromQueryIntegrationTest extends ReadFromQueryIntegrationTestBase {
 
   // tests are from the super-class